Hand reveal workaround fix
Now a player's hand will automatically be shuffled when the effect of a card that they activated in the hand begins resolving, but only if that same card is still in the hand on resolution.

Fixes cases such as the "Enneacraft" monsters where the opponent shouldn't know if the player Special Summoned the monster whose effect was activated or not.