Xyz proc update
Allow the player to finish selection early when using a monster that can be treated as multiple Xyz Materials (and the player has already selected enough materials) for the Xyz Summon of a monster that doesn't have a specific number of materials as the maximum (e.g. "3+ ...").

credits to edo